feature(schema): validate block children for supported object types
Block members can not be given a type that is not a object-like type.

A type like 'string' for instance doesn't make sense inside a block child,
as they always have to be wrapped in a object.

I have identified the following supported core types:

['file', 'image', 'object', 'reference']

It is still possible to use shorthands for hoisted object-like types.
If 'author' is defined as a object elsewhere, {type: 'author'} will
still be allowed.

Everything else in the block types 'of' property
will throw a schema validation error with a listing of supported types.