fix(portable-text-editor): fix issue with fast toggling of decorators
Found a bug where if you toggle different decorators in sequence while
typing it can sometimes reset your applied decorator when reconciling with
the server value (as the mark wasn't correcty set with a patch).

This was happening because we didn't specify the _type when splitting spans
when assigning the new mark. The normalizer function would then assign the
'span' type to it, but not write the marks on the node while doing so.

With this fix the value and patches generated will be correct in the first
place, and it will not need to be normalized like before.