test(playwright-ct): add to tests (#4876)
* test(playwright-ct): improve PT-input decorator test

* test(playwright-ct): simplify assertion

* test(playwright-ct): add missing await statement

* test(playwright-ct): test if overlay is there first

This makes it possible to call these helpers at any time.

* test(playwright-ct): use standard playwright keyboard hotkey activation

After resolving an issue with togglign marks recently, this can now be called the standard way without problems.

* test(playwright-ct): fix issue with flaky webkit tests

Refactor PT-iput overlay activation to reuse code, and get
around an issue on Webkit/Linux where tests would randomly
fail because of the field is not ready before we start
to interact with it.

By adding this:

```ts
await page.waitForSelector(`[data-testid='${testId}']`)`
```

before starting to interact, the problems seems gone.

* test(playwright-ct): remove debug statement

* test(playwright-ct): simplify selectors

Keeping it simple

* test(playwright-ct): decrease expect timout from 20 > 5

* test(playwright-ct): adjust timouts

* test(playwright-ct): ensure visibility of locator

* test(playwright-ct): wait for toolbar to become ready

* test(playwright-ct): add to Decorator test

* test(playwright-ct): use toBeVisible asserts in PT-input placeholder tests

* test(form/inputs): add testid to PT-input toolbar buttons

* test(playwright-ct): fix issue with decorator test on webkit