feat(WCAG): Associate field descriptions to inputs (#4896)
* feat(WCAG): Associate field descriptions to inputs

* feat(WCAG): Adds aria-describedby to PTE

* feat(WCAG): Rename PT prop and change constructDescriptionId for createDescriptionId

* fix(portable-text-editor): forward html attributes to editable component

This will forward html props for the editable element to the slate editable.
It will also get rid of the need of having an own wrapper div element to
put the ref on. Point it directly on the editable element.

* test(portable-text-editor): update snapshot

* refactor(portable-text-editor): remove unnecessary memo

This doesn't need to be memoed, at least not after supporting restProps here.

Debugging shows that the nodes inside are not re-rendered by removing this memo,
which is what we care about.

---------

Co-authored-by: Per-Kristian Nordnes <per.kristian.nordnes@gmail.com>