fix(portable-text-editor): fix issues for .delete fn and add tests (#5060)
* fix(portable-text-editor): ensure that the range is valid before deleting

This will make sure that the produced Range actually is pointing to content. Error if the Range is invalid.

* fix(portable-text-editor): fix issue where all the content is deleted

There was a bug in the delete function that could leave the editor in a broken state
if the last block is removed from the content.

* fix(portable-text-editor): fix issues with delete transforms

After writing tests for this, I found some issues here. Make separate code paths for each mode.

* test(portable-text-editor): write more tests for .delete function